# Conflicts:
#	qt-trading-web/src/main/resources/templates/layout/sidebar.html

+++ qt-trading-web/src/main/java/com/qtdbp/trading/mapper/DataAuthorizeOrderMapper.java
@@ -36,6 +36,6 @@ public interface DataAuthorizeOrderMapper {
      * @param demandOrder
      * @return
      */
-    Map<String, Integer> findAllOrderInfo(DataAuthorizeOrderModel demandOrder);
+    Map<String, Long> findAllOrderInfo(DataAuthorizeOrderModel demandOrder);
 
 }
